Make preprocess and GDP_DATA specific to hourly and 6-hourly modules
